Fix several bugs in PNet implementation.
Fixes a bug in the PNet implementation that would accept process retransmissions instead of discarding them.
This bug was responsible for the "almost-infinite" loop of instantiations ("pony centipede" bug)

Fixes a bug where an unknown message in a packet would cause the whole packet to be discarded.

Fixes a bug where two players leaving/entering at the same time could be spawned twice, or not at all.
WARNING: This bugfix uses a mutex. In case of deadlock, please REPORT IT immediatly and send the server log !