Some errors aren't status messages anymore.
Errors when sending a message aren't status messages anymore, as long as we manage to restart the UDP server.
Add a descriptive error message when we failed to send a message.

Sending a save to a netview ID that doesn't exist isn't a status message anymore. We just ignore the missing player.