Upload optional DLL: ALICORN OC EDITION
Everyone (except mooses) becomes an Alicorn !
Notice : This is just for show. Unicorns still can't use fly and
mudponies still can't do shit. No refund. The author will not be held
responsible if you catch autism.

This is not on by default, to activate it, replace Assembly-CSharp.dll
with this new patched DLL.