Real skills effects, create StatsComponent
We now use a real skill's effects as defined in Skills.json.
Teleport is implemented separately, since it's main effect (teleporting)
is not described in Skills.json
Other "special" effects not described in Skills.json are not implemented

The skill/battle system isn't finished, it supports:
- Teleporting
- DPS and instant damages
- Targeting (None, Friendly, Self, Ennemies)
- Only the target is affected (by the target and splash effects), the
server does not try to find the untargeted entities affected by the
splash at the moment
- Death/respawning (instantly) for all StatsComponents
- Per-entit base defense stat

This commit also reduces coupling of Widget and Skill, fixes DPS being
mispelled as DSP, takes Duration of effects into account, and fixes mobs
not sending their health stat of 0 when dying.